Backend-FastAPI/pyproject.toml
[project]
name = "backend-FastAPI"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro",email = "pedro.felipe11@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
]

[tool.poetry]
packages = [{include = "curso_api", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=curso_API -vv'
post_test = 'coverage html'
migration = 'alembic revision --autogenerate -m'
up_db = 'alembic upgrade head'